requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
contextlib2==21.6.0
cryptography==44.0.0
Django==4.2.19
django-ninja==1.3.0
django-ninja-extra==0.22.3
djangorestframework==3.15.2
injector==0.22.0
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
